Add evalOnUiThread() method, and make runOnUiThread() return Unit.
Some operations on the UI thread need to return values. In order to do so
safely, we need a method of enqueueing a computation behind a container
which supports a blocking wait for the result.

evalOnUiThread() takes a code block just like runOnUiThread(), but
returns a scala.concurrent.Future wrapper for the result. In most cases
this can be dereferenced by doing:

val res = Await.result(f, Duration.Inf)

which will block indefinitely in the current thread, waiting for the
submitted computation to execute on the UI thread. Once the computation
completes, either res will contain the successful result. If an exception
is thrown in the computation, or the it will be re-thrown in the *current*
thread where Await.result() is called.

If Await.result() (or other appropriate handling of the Future) is never
called, the result OR exception is discarded. Make sure to do something
with the Future so an exception result goes somewhere useful!

(By contrast, runOnUiThread() instead lets an exception bubble up the UI
thread's stack since it has nowhere else to send it. This also means that
evalOnUiThread() can be used for a code block which needs to propagate
an exception outside the UI thread but otherwise has no useful result.)

Thanks to PkmX@github for the basic implementation behind evalOnUiThread.

Fixes #118.