Propagate Android API deprecation warnings to scaloid-using code.
There are various Android APIs which have been deprecated over time,
and the deprecation warnings are masked by scaloid. This means that
code using scaloid won't see the deprecation warnings intended by
the Android API libraries.

So here, note the deprecated elements and mark the corresponding
scaloid elements as deprecated. This both silences most of the
compile-time deprecation warnings for scaloid itself, and creates
deprecation warnings in scaloid-using code. No deprecation message
is included (it's just @deprecated("","")).

Note that this makes the generated code dependent on the highest
API level available in $ANDROID_HOME, as that will be used to mark
elements as deprecated. So along with this change, contributor wiki
instructions should be updated to indicate the API level in use by
the scaloid maintainer -- so that scaloid:generate produces the
correct annotations for the newest public Android API.

Along with this change, update some non-generated code with
explicit @deprecated annotations for things deprecated as of API 21.

ClipboardManager is explicitly not deprecated as it changed in
API 11 with a richer, but compatible, subclass of the original.