Use partially applied function syntax to avoid wrapping by-name thunks.
Given a by-name block named "thunk", the syntax

  (() => thunk)

wraps the thunk in an extra Function0 object. However, by-name thunks
are already Function0 instances, so a micro-optimization can be done by
using the syntax

  thunk _

This tells the compiler that the by-name thunk should be partially
applied (retained as an explicit function, not executed). This removes
one layer of call indirection for the on*() callbacks in SActivity and
other Creatable/Destroyable classes.